Move Divine Storm and Zealotry talents to paladin package
