+++ sim/paladin/retribution/retribution.go
@@ -54,8 +54,6 @@ func (ret *RetributionPaladin) Initialize() {
 	ret.Paladin.Initialize()
 	ret.RegisterSpecializationEffects()
 	ret.RegisterTemplarsVerdict()
-	ret.RegisterDivineStorm()
-	ret.RegisterZealotry()
 }
 
 func (ret *RetributionPaladin) ApplyTalents() {
