--- sim/paladin/retribution/zealotry.go
@@ -1,43 +1,0 @@
-package retribution
-
-import (
-	"github.com/wowsims/cata/sim/core"
-	"github.com/wowsims/cata/sim/paladin"
-	"time"
-)
-
-func (retPaladin *RetributionPaladin) RegisterZealotry() {
-	if !retPaladin.Talents.Zealotry {
-		return
-	}
-
-	actionId := core.ActionID{SpellID: 85696}
-
-	retPaladin.ZealotryAura = retPaladin.RegisterAura(core.Aura{
-		Label:    "Zealotry",
-		ActionID: actionId,
-		Duration: 20 * time.Second,
-	})
-
-	retPaladin.Zealotry = retPaladin.RegisterSpell(core.SpellConfig{
-		ActionID:       actionId,
-		Flags:          core.SpellFlagAPL,
-		ClassSpellMask: paladin.SpellMaskZealotry,
-
-		Cast: core.CastConfig{
-			CD: core.Cooldown{
-				Timer:    retPaladin.NewTimer(),
-				Duration: 2 * time.Minute,
-			},
-		},
-		ExtraCastCondition: func(sim *core.Simulation, target *core.Unit) bool {
-			return retPaladin.GetHolyPowerValue() >= 3
-		},
-
-		ThreatMultiplier: 1,
-
-		ApplyEffects: func(sim *core.Simulation, target *core.Unit, spell *core.Spell) {
-			retPaladin.ZealotryAura.Activate(sim)
-		},
-	})
-}
